+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:theme="@style/Theme.Android_repo_05"
         tools:targetApi="31">
         <activity
-            android:name=".ui.activities.LoginActivity"
+            android:name=".ui.login.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -39,14 +39,14 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.activities.SearchActivity"
+            android:name=".ui.search.SearchActivity"
             android:exported="false"
-            android:windowSoftInputMode="stateAlwaysVisible"/>
+            android:windowSoftInputMode="stateAlwaysVisible" />
         <activity
-            android:name=".ui.activities.ProfileActivity"
+            android:name=".ui.profile.ProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.activities.MainActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="false" />
     </application>
