--- app/src/main/java/com/example/android_repo_05/ui/viewmodels/TokenViewModel.kt
@@ -1,24 +1,0 @@
-package com.example.android_repo_05.ui.viewmodels
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.example.android_repo_05.data.models.ResponseState
-import com.example.android_repo_05.data.models.TokenModel
-import com.example.android_repo_05.data.repositories.TokenRepository
-import kotlinx.coroutines.launch
-
-class TokenViewModel(private val repository: TokenRepository) : ViewModel() {
-    private var _tokenModel: MutableLiveData<ResponseState<TokenModel>> = MutableLiveData()
-    val tokenModel: LiveData<ResponseState<TokenModel>> get() = _tokenModel
-
-    fun getAccessTokenFromRemote(code: String) = viewModelScope.launch {
-        _tokenModel.postValue(ResponseState.Loading())
-        _tokenModel.postValue(repository.getAccessTokenFromRemote(code))
-    }
-
-    fun setAccessTokenToDataStore(accessToken: String) = viewModelScope.launch {
-        repository.setAccessToken(accessToken = accessToken)
-    }
-}
