+++ app/src/main/res/layout/activity_main.xml
@@ -10,7 +10,7 @@
             type="String" />
         <variable
             name="mainActivity"
-            type="com.example.android_repo_05.ui.activities.MainActivity" />
+            type="com.example.android_repo_05.ui.main.MainActivity" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -48,7 +48,7 @@
             android:layout_marginHorizontal="24dp"
             app:layout_constraintTop_toBottomOf="@id/l_tb_main">
 
-            <com.example.android_repo_05.customviews.MainTabButton
+            <com.example.android_repo_05.ui.main.custom.MainTabButton
                 android:id="@+id/tab_btn_issue"
                 android:layout_width="0dp"
                 android:layout_height="match_parent"
@@ -59,7 +59,7 @@
                 app:layout_constraintEnd_toStartOf="@id/tab_btn_notification"
                 app:layout_constraintStart_toStartOf="parent" />
 
-            <com.example.android_repo_05.customviews.MainTabButton
+            <com.example.android_repo_05.ui.main.custom.MainTabButton
                 android:id="@+id/tab_btn_notification"
                 android:layout_width="0dp"
                 android:layout_height="match_parent"
